inference: use the same method lookup cache across same inference trial
Previously the method lookup result was created per frame and so the
look cache hasn't been use that much. With this change the cache is
created per inference, and so the cached result will be used when we
already saw the same match in the same inference shot, and it may speed
up the lookup time a bit.

This commit also setups new `AbstractInterpreter` interface `get_method_lookup_cache`
which specifies what method lookup cache is used by each `AbstractInterpreter`.
`NativeInterpreter` creates a cache per inference, and so it is valid
since lookup is done in the same world age in the same inference shot.
External `AbstractInterpreter` doesn't opt into this cache by default,
and its behavior won't change in anyway.